-refactor the details screen code - avoid the null image error - test the share movie
