+++ lib/core/routing/app_router.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:movie_app/core/routing/routes.dart';
-import 'package:movie_app/features/details/movie_details.dart';
-import 'package:movie_app/features/details/ui/detail_screen.dart';
+import 'package:movie_app/features/details/ui/movie_details.dart';
 import 'package:movie_app/features/home_screen/ui/home_screen.dart';
 import 'package:movie_app/features/login/logic/login_cubit.dart';
 import 'package:movie_app/features/on_boarding/ui/screens/on_boarding_screen.dart';
