+++ lib/features/details/ui/widgets/custom__icon_button.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/cupertino.dart';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
@@ -7,30 +7,34 @@ class CustomIconButton extends StatelessWidget {
     super.key,
     required this.icon,
     required this.onTap,
-    required this.color,
-    this.value = 5,
+    this.color,
     this.iconColor = Colors.white,
+    this.radius,
+    this.iconSize,
+    this.leftPadding,
   });
   final IconData icon;
   final void Function() onTap;
 
-  final Color color;
-  final double value;
-  final iconColor ;
+  final Color? color;
+  final double? radius;
+  final double? iconSize;
+  final double? leftPadding;
+  final iconColor;
   @override
   Widget build(BuildContext context) {
     return InkWell(
       onTap: onTap,
-      child: Container(
-        padding: EdgeInsets.all(value),
-        decoration: BoxDecoration(
-          shape: BoxShape.circle,
-          color: color,
-        ),
-        child: Icon(
-          icon,
-          color: iconColor,
-          size: 27.sp,
+      child: CircleAvatar(
+        backgroundColor: color ?? Colors.black38,
+        radius: radius ?? 20.sp,
+        child: Padding(
+          padding: EdgeInsets.only(left: leftPadding ?? 0.0),
+          child: Icon(
+            icon,
+            color: iconColor,
+            size: iconSize ?? 27.sp,
+          ),
         ),
       ),
     );
