+++ lib/features/home_screen/ui/widgets/shimmer_trending_list_view_item.dart
@@ -22,7 +22,7 @@ class ShimmerTrendingListViewItem extends StatelessWidget {
                   baseColor: Colors.grey[300]!,
                   highlightColor: Colors.grey[100]!,
                   child: Container(
-                    height: 260,
+                    height: 280,
                     width: double.infinity,
                     color: Colors.white,
                   ),
